@@ -1,21 +1,18 @@
 module github.com/tomochain/tomochain
 
-go 1.13
+go 1.19
 
 require (
 	bazil.org/fuse v0.0.0-20180421153158-65cc252bf669
-	github.com/aead/siphash v1.0.1 // indirect
 	github.com/VictoriaMetrics/fastcache v1.5.7
 	github.com/aristanetworks/goarista v0.0.0-20191023202215-f096da5361bb
 	github.com/btcsuite/btcd v0.0.0-20171128150713-2e60448ffcc6
-	github.com/btcsuite/winsvc v1.0.0 // indirect
 	github.com/cespare/cp v1.1.1
 	github.com/davecgh/go-spew v1.1.1
 	github.com/deckarep/golang-set v0.0.0-20180603214616-504e848d77ea
 	github.com/docker/docker v1.4.2-0.20180625184442-8e610b2b55bf
+	github.com/dop251/goja v0.0.0-20230531210528-d7324b2d74f7
 	github.com/edsrzf/mmap-go v1.0.0
-	github.com/elastic/gosigar v0.10.5
-	github.com/ethereum/go-ethereum v1.9.9
 	github.com/fatih/color v1.6.0
 	github.com/gizak/termui v2.2.0+incompatible
 	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8
