@@ -17,46 +17,41 @@
 package jsre
 
 import (
-	"io/ioutil"
 	"os"
 	"path"
+	"reflect"
 	"testing"
 	"time"
 
-	"github.com/robertkrimen/otto"
+	"github.com/dop251/goja"
 )
 
-type testNativeObjectBinding struct{}
+type testNativeObjectBinding struct {
+	vm *goja.Runtime
+}
 
 type msg struct {
 	Msg string
 }
 
-func (no *testNativeObjectBinding) TestMethod(call otto.FunctionCall) otto.Value {
-	m, err := call.Argument(0).ToString()
-	if err != nil {
-		return otto.UndefinedValue()
-	}
-	v, _ := call.Otto.ToValue(&msg{m})
-	return v
+func (no *testNativeObjectBinding) TestMethod(call goja.FunctionCall) goja.Value {
+	m := call.Argument(0).ToString().String()
+	return no.vm.ToValue(&msg{m})
 }
 
-func newWithTestJS(t *testing.T, testjs string) (*JSRE, string) {
-	dir, err := ioutil.TempDir("", "jsre-test")
-	if err != nil {
-		t.Fatal("cannot create temporary directory:", err)
-	}
+func newWithTestJS(t *testing.T, testjs string) *JSRE {
+	dir := t.TempDir()
 	if testjs != "" {
-		if err := ioutil.WriteFile(path.Join(dir, "test.js"), []byte(testjs), os.ModePerm); err != nil {
+		if err := os.WriteFile(path.Join(dir, "test.js"), []byte(testjs), os.ModePerm); err != nil {
 			t.Fatal("cannot create test.js:", err)
 		}
 	}
-	return New(dir, os.Stdout), dir
+	jsre := New(dir, os.Stdout)
+	return jsre
 }
 
 func TestExec(t *testing.T) {
-	jsre, dir := newWithTestJS(t, `msg = "testMsg"`)
-	defer os.RemoveAll(dir)
+	jsre := newWithTestJS(t, `msg = "testMsg"`)
 
 	err := jsre.Exec("test.js")
 	if err != nil {
@@ -66,11 +61,11 @@ func TestExec(t *testing.T) {
 	if err != nil {
 		t.Errorf("expected no error, got %v", err)
 	}
-	if !val.IsString() {
+	if val.ExportType().Kind() != reflect.String {
 		t.Errorf("expected string value, got %v", val)
 	}
 	exp := "testMsg"
-	got, _ := val.ToString()
+	got := val.ToString().String()
 	if exp != got {
 		t.Errorf("expected '%v', got '%v'", exp, got)
 	}
@@ -78,25 +73,24 @@ func TestExec(t *testing.T) {
 }
 
 func TestNatto(t *testing.T) {
-	jsre, dir := newWithTestJS(t, `setTimeout(function(){msg = "testMsg"}, 1);`)
-	defer os.RemoveAll(dir)
+	jsre := newWithTestJS(t, `setTimeout(function(){msg = "testMsg"}, 1);`)
 
 	err := jsre.Exec("test.js")
 	if err != nil {
-		t.Errorf("expected no error, got %v", err)
+		t.Fatalf("expected no error, got %v", err)
 	}
 	time.Sleep(100 * time.Millisecond)
 	val, err := jsre.Run("msg")
 	if err != nil {
-		t.Errorf("expected no error, got %v", err)
+		t.Fatalf("expected no error, got %v", err)
 	}
-	if !val.IsString() {
-		t.Errorf("expected string value, got %v", val)
+	if val.ExportType().Kind() != reflect.String {
+		t.Fatalf("expected string value, got %v", val)
 	}
 	exp := "testMsg"
-	got, _ := val.ToString()
+	got := val.ToString().String()
 	if exp != got {
-		t.Errorf("expected '%v', got '%v'", exp, got)
+		t.Fatalf("expected '%v', got '%v'", exp, got)
 	}
 	jsre.Stop(false)
 }
@@ -105,7 +99,7 @@ func TestBind(t *testing.T) {
 	jsre := New("", os.Stdout)
 	defer jsre.Stop(false)
 
-	jsre.Bind("no", &testNativeObjectBinding{})
+	jsre.Set("no", &testNativeObjectBinding{vm: jsre.vm})
 
 	_, err := jsre.Run(`no.TestMethod("testMsg")`)
 	if err != nil {
@@ -114,8 +108,7 @@ func TestBind(t *testing.T) {
 }
 
 func TestLoadScript(t *testing.T) {
-	jsre, dir := newWithTestJS(t, `msg = "testMsg"`)
-	defer os.RemoveAll(dir)
+	jsre := newWithTestJS(t, `msg = "testMsg"`)
 
 	_, err := jsre.Run(`loadScript("test.js")`)
 	if err != nil {
@@ -125,11 +118,11 @@ func TestLoadScript(t *testing.T) {
 	if err != nil {
 		t.Errorf("expected no error, got %v", err)
 	}
-	if !val.IsString() {
+	if val.ExportType().Kind() != reflect.String {
 		t.Errorf("expected string value, got %v", val)
 	}
 	exp := "testMsg"
-	got, _ := val.ToString()
+	got := val.ToString().String()
 	if exp != got {
 		t.Errorf("expected '%v', got '%v'", exp, got)
 	}
