@@ -17,5 +17,12 @@
 // Package deps contains the console JavaScript dependencies Go embedded.
 package deps
 
-//go:generate go-bindata -nometadata -pkg deps -o bindata.go bignumber.js web3.js
-//go:generate gofmt -w -s bindata.go
+import (
+	_ "embed"
+)
+
+//go:embed web3.js
+var Web3JS string
+
+//go:embed bignumber.js
+var BigNumberJS string
